pyproject.toml
[project]
description = "mobile01 特定文章爬取"
name = "mobile01-scrape"
requires-python = ">=3.11"

[tool.poetry]
authors = ["seria <seria.ati@gmail.com>"]
description = "mobile01 特定文章爬取"
license = "GPL-3.0"
name = "mobile01-scrape"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
loguru = "^0.7.2"
playwright = "^1.45.1"
pydantic = "^2.7.2"
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
